+++ sdk/storagemover/arm-storagemover/samples/v3/typescript/src/connectionsGetSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to gets a Connection resource.
  *
  * @summary gets a Connection resource.
- * x-ms-original-file: 2025-12-01/Connections_Get.json
+ * x-ms-original-file: 2026-05-01/Connections_Get.json
  */
 async function connectionsGet(): Promise<void> {
   const credential = new DefaultAzureCredential();