+++ sdk/storagemover/arm-storagemover/samples/v3/typescript/src/jobDefinitionsUpdateSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to updates properties for a Job Definition resource. Properties not specified in the request body will be unchanged.
  *
  * @summary updates properties for a Job Definition resource. Properties not specified in the request body will be unchanged.
- * x-ms-original-file: 2025-12-01/JobDefinitions_Update.json
+ * x-ms-original-file: 2026-05-01/JobDefinitions_Update.json
  */
 async function jobDefinitionsUpdate(): Promise<void> {
   const credential = new DefaultAzureCredential();